gcs sdk tests
